corrections .conf

+++ srcs/requirements/wordpress/wp_conf.sh
@@ -1,43 +1,48 @@
 #!/bin/bash
 
-set -e
-
 #-----------------------INSTALL WP-CLI
-curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
-chmod +x wp-cli.phar
-mv wp-cli.phar /usr/bin/wp
+if [ ! -f /var/www/wordpress/wp-config.php ]; then
+    echo "Installation de Wordpress..."
 
-#-----------------------PREPARATION WORDPRESS DIRECTORY
-mkdir -p /var/www/wordpress
-cd /var/www/wordpress
-chmod -R 755 /var/www/wordpress
-chown -R www-data:www-data /var/www/wordpress
+    curl -O https://raw.githubusercontent.com/wp-cli/builds/gh-pages/phar/wp-cli.phar
+    chmod +x wp-cli.phar
+    mv wp-cli.phar /usr/bin/wp
 
-#-----------------------INSTALL WORDPRESS
-wp core download --allow-root
+    #-----------------------PREPARATION WORDPRESS DIRECTORY
+    mkdir -p /var/www/wordpress
+    cd /var/www/wordpress
+    chmod -R 755 /var/www/wordpress
+    chown -R www-data:www-data /var/www/wordpress
 
-# configuration wp-config.php 
-wp core config \
-    --dbhost="mariadb:3306" \
-    --dbname="$SQL_DATABASE" \
-    --dbuser="$SQL_USER" \
-    --dbpass="$SQL_PWD" \
-    --allow-root 
+    #-----------------------INSTALL WORDPRESS
+    wp core download --allow-root
 
-# wordpress installation + configuration
-wp core install \
-    --url="$WP_DOMAIN_NAME" \
-    --title="$WP_TITLE" \
-    --admin_user="$WP_ADMIN_NAME" \
-    --admin_password="$WP_ADMIN_PWD" \
-    --admin_email="$WP_ADMIN_EMAIL" \
-    --allow-root
+    # configuration wp-config.php 
+    wp core config \
+        --dbhost="mariadb:3306" \
+        --dbname="$SQL_DATABASE" \
+        --dbuser="$SQL_USER" \
+        --dbpass="$SQL_PWD" \
+        --allow-root 
 
-# create a new user
-wp user create "$WP_USER_NAME" "$WP_USER_EMAIL" \
-    --user-pass="$WP_USER_PWD" \
-    --role="$WP_USER_ROLE" \
-    --allow-root
+    # wordpress installation + configuration
+    wp core install \
+        --url="$WP_DOMAIN_NAME" \
+        --title="$WP_TITLE" \
+        --admin_user="$WP_ADMIN_NAME" \
+        --admin_password="$WP_ADMIN_PWD" \
+        --admin_email="$WP_ADMIN_EMAIL" \
+        --allow-root
+
+    # create a new user
+    wp user create "$WP_USER_NAME" "$WP_USER_EMAIL" \
+        --user-pass="$WP_USER_PWD" \
+        --role="$WP_USER_ROLE" \
+        --allow-root
+    echo "Wordpress installe avec succes !"
+else
+    echo "Wordpress deja configure, damarrage..."
+fi
 
 #-----------------------PHP-FPM CONFIGURATION
 # modificate line 36 of www.conf to listen = 9000
